Fixed inheritence in Site
(constructor of virtual base class)

+++ src/Bpp/Seq/SymbolList.h
@@ -156,7 +156,7 @@ public:
 public:
   std::shared_ptr<const Alphabet> getAlphabet() const override { return alphabet_; }
 
-  size_t size() const override { return static_cast<size_t>(content_.size()); }
+  size_t size() const override { return content_.size(); }
 
   virtual void setContent(const std::vector<T>& list) override
   {
